requirements.txt
aiohttp==3.7.4.post0
discord.py==1.7.3
ython-dotenv==0.18.0
